+++ cmd/backup.go
@@ -106,6 +106,7 @@ func RunBackup(ctx context.Context, flagMap map[string]any) error {
 		),
 		pathcompression.NewPathCompressor(
 			runConfig.Engine.Performance.BufferSizeKB,
+			runConfig.Engine.Performance.ReadAheadLimitKB,
 			runConfig.Engine.Performance.CompressWorkers,
 		),
 	)