+++ cmd/restore.go
@@ -102,6 +102,7 @@ func RunRestore(ctx context.Context, flagMap map[string]any) error {
 		),
 		pathcompression.NewPathCompressor(
 			runConfig.Engine.Performance.BufferSizeKB,
+			runConfig.Engine.Performance.ReadAheadLimitKB,
 			runConfig.Engine.Performance.CompressWorkers,
 		),
 	)